requirements.txt
django
djangorestframework
psycopg2-binary
langchain
langgraph
ollama
streamlit
requests
redis
psycopg2-binary